Split combigrid in header/source files
