+++ combigrid/src/sgpp/combigrid/grid/IndexVectorIterator.hpp
@@ -18,11 +18,11 @@ namespace combigrid {
 class IndexVectorIterator : public std::iterator<std::random_access_iterator_tag, IndexVector,
                                                  size_t, IndexVector*, IndexVector&> {
  public:
-  inline IndexVectorIterator() : dim(0), indexVector(), minIndex(), maxIndex(),
+  IndexVectorIterator() : dim(0), indexVector(), minIndex(), maxIndex(),
       numberOfIndexVectors(0), sequenceIndex(0) {
   }
 
-  inline explicit IndexVectorIterator(const FullGrid& grid) : dim(grid.getDimension()),
+  explicit IndexVectorIterator(const FullGrid& grid) : dim(grid.getDimension()),
       indexVector(dim), minIndex(dim), maxIndex(dim),
       numberOfIndexVectors(dim), sequenceIndex(0) {
     grid.getMinIndex(minIndex);
@@ -30,7 +30,7 @@ class IndexVectorIterator : public std::iterator<std::random_access_iterator_tag
     grid.getNumberOfIndexVectors(numberOfIndexVectors);
   }
 
-  inline IndexVectorIterator(const IndexVector& minIndex, const IndexVector& maxIndex) :
+  IndexVectorIterator(const IndexVector& minIndex, const IndexVector& maxIndex) :
       dim(minIndex.size()), indexVector(dim), minIndex(minIndex), maxIndex(maxIndex),
       numberOfIndexVectors(dim), sequenceIndex(0) {
     for (size_t d = 0; d < dim; d++) {
@@ -38,26 +38,26 @@ class IndexVectorIterator : public std::iterator<std::random_access_iterator_tag
     }
   }
 
-  inline IndexVectorIterator(const IndexVectorIterator&) = default;
-  inline IndexVectorIterator& operator=(const IndexVectorIterator&) = default;
+  IndexVectorIterator(const IndexVectorIterator&) = default;
+  IndexVectorIterator& operator=(const IndexVectorIterator&) = default;
 
-  inline size_t getSequenceIndex() const {
+  size_t getSequenceIndex() const {
     return sequenceIndex;
   }
 
-  inline void setSequenceIndex(size_t sequenceIndex) {
+  void setSequenceIndex(size_t sequenceIndex) {
     this->sequenceIndex = sequenceIndex;
   }
 
-  inline IndexVector& operator*() {
+  IndexVector& operator*() {
     return operator[](sequenceIndex);
   }
 
-  inline IndexVector* operator->() {
+  IndexVector* operator->() {
     return &operator[](sequenceIndex);
   }
 
-  inline IndexVector& operator[](size_t rhs) {
+  IndexVector& operator[](size_t rhs) {
     for (size_t d = 0; d < dim; d++) {
       indexVector[d] = minIndex[d] + static_cast<index_t>(rhs % numberOfIndexVectors[d]);
       rhs /= numberOfIndexVectors[d];
@@ -66,81 +66,81 @@ class IndexVectorIterator : public std::iterator<std::random_access_iterator_tag
     return indexVector;
   }
 
-  inline IndexVectorIterator& operator++() {
+  IndexVectorIterator& operator++() {
     sequenceIndex++;
     return *this;
   }
 
-  inline IndexVectorIterator operator++(int) {
+  IndexVectorIterator operator++(int) {
     IndexVectorIterator tmp(*this);
     sequenceIndex++;
     return tmp;
   }
 
-  inline IndexVectorIterator& operator--() {
+  IndexVectorIterator& operator--() {
     sequenceIndex--;
     return *this;
   }
 
-  inline IndexVectorIterator operator--(int) {
+  IndexVectorIterator operator--(int) {
     IndexVectorIterator tmp(*this);
     sequenceIndex--;
     return tmp;
   }
 
-  inline IndexVectorIterator operator+(size_t rhs) const {
+  IndexVectorIterator operator+(size_t rhs) const {
     IndexVectorIterator tmp(*this);
     tmp.sequenceIndex += rhs;
     return tmp;
   }
 
-  friend inline IndexVectorIterator operator+(size_t lhs, const IndexVectorIterator& rhs) {
+  friend IndexVectorIterator operator+(size_t lhs, const IndexVectorIterator& rhs) {
     IndexVectorIterator tmp(rhs);
     tmp.sequenceIndex += lhs;
     return tmp;
   }
 
-  inline IndexVectorIterator operator-(size_t rhs) const {
+  IndexVectorIterator operator-(size_t rhs) const {
     IndexVectorIterator tmp(*this);
     tmp.sequenceIndex -= rhs;
     return tmp;
   }
 
-  inline size_t operator-(const IndexVectorIterator& other) const {
+  size_t operator-(const IndexVectorIterator& other) const {
     return sequenceIndex - other.sequenceIndex;
   }
 
-  inline IndexVectorIterator& operator+=(size_t rhs) {
+  IndexVectorIterator& operator+=(size_t rhs) {
     sequenceIndex += rhs;
     return *this;
   }
 
-  inline IndexVectorIterator& operator-=(size_t rhs) {
+  IndexVectorIterator& operator-=(size_t rhs) {
     sequenceIndex -= rhs;
     return *this;
   }
 
-  inline bool operator==(const IndexVectorIterator& other) const {
+  bool operator==(const IndexVectorIterator& other) const {
     return (sequenceIndex == other.sequenceIndex);
   }
 
-  inline bool operator!=(const IndexVectorIterator& other) const {
+  bool operator!=(const IndexVectorIterator& other) const {
     return (sequenceIndex != other.sequenceIndex);
   }
 
-  inline bool operator<(const IndexVectorIterator& other) const {
+  bool operator<(const IndexVectorIterator& other) const {
     return (sequenceIndex < other.sequenceIndex);
   }
 
-  inline bool operator>(const IndexVectorIterator& other) const {
+  bool operator>(const IndexVectorIterator& other) const {
     return (sequenceIndex > other.sequenceIndex);
   }
 
-  inline bool operator<=(const IndexVectorIterator& other) const {
+  bool operator<=(const IndexVectorIterator& other) const {
     return (sequenceIndex <= other.sequenceIndex);
   }
 
-  inline bool operator>=(const IndexVectorIterator& other) const {
+  bool operator>=(const IndexVectorIterator& other) const {
     return (sequenceIndex >= other.sequenceIndex);
   }
