+++ combigrid/src/sgpp/combigrid/operation/OperationPole.hpp
@@ -9,8 +9,6 @@
 #include <sgpp/base/datatypes/DataVector.hpp>
 #include <sgpp/combigrid/LevelIndexTypes.hpp>
 
-#include <memory>
-
 namespace sgpp {
 namespace combigrid {
